// Angular/src/app/modules/user-profile/personal-information/model/personal-information.model.ts
export class PersonalInfoListModel {
    id: number;
    first_name: any;
    last_name: any;
    postal_code: any;
    joining_date: any;
    groups: any;
    email: any;
    employee_code: any;
    reporting_manager_name: any;
    street_address: any;
    designation: any;
    department: any;
    country: any;
    state: any;
    city: any;
    band: any;
    birthday: any;
    team : any;
    phone : any;
    profile_image: any;
    thumbnail_photo_base64: any;
    status: any;
  }